requirements.txt
python-dotenv
python-box
PyYAML
ipykernel
numpy 
pandas 
matplotlib 
seaborn 
scikit-learn
mlflow
dvc
dvc-s3
torch 
skorch
gradio